remove commands
