--- commands/tex_template/latex.tex
@@ -1,28 +1,0 @@
-\documentclass[uplatex]{jsarticle}
-
-\usepackage{amsmath,amssymb}
-\usepackage{ascmac}
-\usepackage{enumerate}
-\usepackage{mathrsfs}
-\usepackage{bm}
-\usepackage{chemfig}
-\usepackage[version=3]{mhchem}
-\usepackage{tikz}
-\usepackage{pxrubrica}
-\usepackage{siunitx}
-\usepackage{circuitikz}
-\usepackage{cancel}
-\usepackage{fancybox}
-\usepackage[top=5mm, left=50mm, right=50mm, bottom=5mm]{geometry}
-
-\pagestyle{empty}
-
-\begin{document}
-\begin{flushright}
-    .
-\end{flushright}
-[REPLACE]
-\begin{flushleft}
-    .
-\end{flushleft}
-\end{document}
