--- commands/tex_template/tex.tex
@@ -1,23 +1,0 @@
-\documentclass[uplatex]{jsarticle}
-
-\usepackage{amsmath,amssymb}
-\usepackage{ascmac}
-\usepackage{enumerate}
-\usepackage{mathrsfs}
-\usepackage{bm}
-\usepackage{chemfig}
-\usepackage[version=3]{mhchem}
-\usepackage{tikz}
-\usepackage{pxrubrica}
-\usepackage{siunitx}
-\usepackage{circuitikz}
-\usepackage{cancel}
-\usepackage{fancybox}
-
-\pagestyle{empty}
-
-\begin{document}
-\begin{gather*}
-[REPLACE]
-\end{gather*}
-\end{document}
